Breaking up the files a bit
Seperating assembly selection window from the table it uses, in preparation for addin a new window/table to select areas
