+++ Editor/UI/TreeViews/AssemblyIdentifier.cs
