+++ Editor/UI/TreeViews/AssemblyTable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1759f7614496041218cf5d4b646631e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
